refactor(core): simplify HTTP server layer
Use Object.fromEntries for query/headers/form parsing, replace duck-
typing with instanceof VertzException, use BadRequestException for
malformed JSON, remove redundant type re-exports from implementation.
