+++ packages/core/src/server/index.ts
@@ -2,4 +2,4 @@ export { parseRequest, parseBody } from './request-utils';
 export type { ParsedRequest } from './request-utils';
 export { createJsonResponse, createErrorResponse } from './response-utils';
 export { handleCors, applyCorsHeaders } from './cors';
-export type { CorsConfig } from './cors';
+export type { CorsConfig } from '../types/app';
