+++ packages/core/src/server/response-utils.ts
@@ -1,3 +1,5 @@
+import { VertzException } from '../exceptions';
+
 export function createJsonResponse(data: unknown, status = 200, headers?: Record<string, string>): Response {
   return new Response(JSON.stringify(data), {
     status,
@@ -8,23 +10,8 @@ export function createJsonResponse(data: unknown, status = 200, headers?: Record
   });
 }
 
-interface VertzLike {
-  statusCode: number;
-  toJSON: () => Record<string, unknown>;
-}
-
-function isVertzException(error: unknown): error is VertzLike {
-  return (
-    typeof error === 'object' &&
-    error !== null &&
-    'statusCode' in error &&
-    'toJSON' in error &&
-    typeof (error as any).toJSON === 'function'
-  );
-}
-
 export function createErrorResponse(error: unknown): Response {
-  if (isVertzException(error)) {
+  if (error instanceof VertzException) {
     return createJsonResponse(error.toJSON(), error.statusCode);
   }
